LPS-52121 Determine mime type by file content and correct extension for set of mime types
Fixing the extension is important for any files that are content disposition inline and especially for swf as it can allow a CSRF attack. Additional mime types we want to fix the extension for can be added to the list. Mime type provided  by the API user can't be trusted as otherwise you'd be able to upload a swf file as a png and thus allow the CSRF attack.